+++ src/dev/mod.rs
@@ -7,6 +7,8 @@ pub mod max7321;
 pub mod pca9536;
 pub mod pca9538;
 pub mod pca9555;
+pub mod pcal6408a;
+pub mod pcal6416a;
 pub mod pcf8574;
 pub mod pcf8575;
 pub mod tca6408a;